+++ pages/api/v1/applicants/[id].ts
@@ -1,4 +1,3 @@
-import { string } from 'joi';
 import { ObjectId } from 'mongodb';
 import { NextApiHandler } from 'next';
 import { SingleApplicantApiResponse } from '../../../../common/types';
@@ -6,6 +5,11 @@ import { connectToDatabase } from '../../../../server/mongoDB';
 import { isAdmin, protect } from '../../../../server/protect';
 
 const applicantHandler: NextApiHandler = async (req, res) => {
+  const admin = await isAdmin(req);
+  if (!admin) {
+    return res.status(401).send({ message: 'User is not an admin' });
+  }
+
   switch (req.method) {
     // get a single applicant
     case 'GET':
@@ -21,13 +25,13 @@ const applicantHandler: NextApiHandler = async (req, res) => {
 };
 
 const getApplicant: NextApiHandler = async (req, res) => {
-  const admin = await isAdmin(req);
-  if (!admin) {
-    return res.status(401).send({ message: 'User is not an admin' });
-  }
   const { userDataCollection } = await connectToDatabase();
   const id = req.query.id;
-  const { _id, ...user } = (await userDataCollection.findOne({ _id: id }))!;
+  const userWithId = await userDataCollection.findOne({ _id: id });
+  if (!userWithId) {
+    return res.status(200).send(null);
+  }
+  const { _id, ...user } = userWithId;
   const body: SingleApplicantApiResponse = {
     user: { _id: _id.toString(), ...user },
   };
@@ -35,15 +39,9 @@ const getApplicant: NextApiHandler = async (req, res) => {
 };
 
 const postApplicant: NextApiHandler = async (req, res) => {
-  const admin = await isAdmin(req);
-  if (!admin) {
-    return res.status(401).send({ message: 'User is not an admin' });
-  }
   const { _id, ...updatedUser } = req.body;
   const { userDataCollection } = await connectToDatabase();
-
   await userDataCollection.replaceOne({ _id: new ObjectId(_id) }, updatedUser);
-
   return res.status(201).send('Successfully updated application or rsvp status.');
 };
